+++ franka_example_controllers/src/cartesian_velocity_example_controller.cpp
@@ -54,8 +54,9 @@ bool CartesianVelocityExampleController::init(hardware_interface::RobotHW* robot
       if (std::abs(state_handle.getRobotState().q_d[i] - q_start[i]) > 0.1) {
         ROS_ERROR_STREAM(
             "CartesianVelocityExampleController: Robot is not in the expected starting position "
-            "for running this example. Run `roslaunch panda_moveit_config move_to_start.launch "
-            "robot_ip:=<robot-ip> load_gripper:=<has-attached-gripper>` first.");
+            "for running this example. Run `roslaunch franka_example_controllers "
+            "move_to_start.launch robot_ip:=<robot-ip> load_gripper:=<has-attached-gripper>` "
+            "first.");
         return false;
       }
     }