+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Requires `libfranka` >= 0.3.0
   * Fixed linker errors when building with `-DFranka_DIR` while an older version of
     `ros-kinetic-libfranka` is installed
   * Added gripper joint state publisher to `franka_visualization`
+  * Moved `move_to_start.py` example script to `franka_example_controllers`
 
 ## 0.2.2 - 2018-01-31